+++ RPGEconomy.Application/Abstractions/Services/IProductTypeService.cs
@@ -7,7 +7,7 @@ public interface IProductTypeService
 {
     Task<Result<ProductTypeDto>> GetByIdAsync(int id);
     Task<Result<IReadOnlyList<ProductTypeDto>>> GetAllAsync();
-    Task<Result<ProductTypeDto>> CreateAsync(string name, string description, double basePrice, double weightPerUnit);
-    Task<Result<ProductTypeDto>> UpdateAsync(int id, string name, string description, double basePrice, double weightPerUnit);
+    Task<Result<ProductTypeDto>> CreateAsync(string name, string description, decimal basePrice, double weightPerUnit);
+    Task<Result<ProductTypeDto>> UpdateAsync(int id, string name, string description, decimal basePrice, double weightPerUnit);
     Task<Result> DeleteAsync(int id);
 }
